move UserBar in a separate component
