+++ src/components/ui/AppButton/AppButton.module.css
@@ -33,12 +33,6 @@
   cursor: not-allowed;
 }
 
-/* .init {
-  border: 1px solid var(--color-scheme-2-border);
-  background: var(--opacity-white-10);
-  color: var(--color-white);
-} */
-
 /* blue  */
 
 .blue {
