+++ src/components/ui/UserBar/UserBar.module.css
@@ -1,0 +1,16 @@
+.userBar {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 8px;
+}
+
+.avatar {
+  width: 32px;
+  height: 32px;
+  border-radius: 50%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: #999;
+}
